Version 1.1 开始搞评论功能

+++ HOA-common/src/main/java/context/UserContext.java
@@ -2,13 +2,13 @@ package context;
 
 public class UserContext {
 
-    public static ThreadLocal<Long> threadLocal = new ThreadLocal<>();
+    public static ThreadLocal<Integer> threadLocal = new ThreadLocal<>();
 
-    public static void setCurrentId(long id) {
+    public static void setCurrentId(int id) {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
